Split runner.go into smaller focused files
Refactor pkg/runner/runner.go (1522 lines) into:
- runner.go (649 lines): Core Runner struct, constructors, orchestration
- execution.go (617 lines): Step execution logic (runStep, runRunStep, etc.)
- volumes.go (152 lines): Volume resolution and mounting logic
- cleanup.go (74 lines): Cleanup and signal handling
- services.go (67 lines): Exposed service tracking

This separation improves code organization and maintainability while
keeping all tests passing.
